Moved Enzo-problem initializers to Enzo/initial
