+++ src/Enzo/_enzo.hpp
@@ -184,30 +184,30 @@ extern "C" {
 
 #include "enzo_EnzoBoundary.hpp"
 
-#include "enzo_EnzoInitialBCenter.hpp"
-#include "enzo_EnzoInitialCloud.hpp"
-#include "enzo_EnzoInitialCollapse.hpp"
-#include "enzo_EnzoInitialCosmology.hpp"
-#include "enzo_EnzoInitialFeedbackTest.hpp"
-#include "enzo_EnzoInitialGrackleTest.hpp"
-#include "enzo_EnzoInitialHdf5.hpp"
-#include "enzo_EnzoInitialImplosion2.hpp"
-#include "enzo_EnzoInitialInclinedWave.hpp"
-#include "enzo_EnzoInitialMusic.hpp"
-#include "enzo_EnzoInitialPm.hpp"
-#include "enzo_EnzoInitialPpmlTest.hpp"
-#include "enzo_EnzoInitialSedovArray2.hpp"
-#include "enzo_EnzoInitialSedovArray3.hpp"
-#include "enzo_EnzoInitialSedovRandom.hpp"
-#include "enzo_EnzoInitialShockTube.hpp"
-#include "enzo_EnzoInitialSoup.hpp"
-#include "enzo_EnzoInitialTurbulence.hpp"
-#include "enzo_EnzoInitialIsolatedGalaxy.hpp"
-#include "enzo_EnzoInitialBurkertBodenheimer.hpp"
-#include "enzo_EnzoInitialMergeSinksTest.hpp"
-#include "enzo_EnzoInitialAccretionTest.hpp"
-#include "enzo_EnzoInitialShuCollapse.hpp"
-#include "enzo_EnzoInitialBBTest.hpp"
+#include "initial/EnzoInitialBCenter.hpp"
+#include "initial/EnzoInitialCloud.hpp"
+#include "initial/EnzoInitialCollapse.hpp"
+#include "initial/EnzoInitialCosmology.hpp"
+#include "initial/EnzoInitialFeedbackTest.hpp"
+#include "initial/EnzoInitialGrackleTest.hpp"
+#include "initial/EnzoInitialHdf5.hpp"
+#include "initial/EnzoInitialImplosion2.hpp"
+#include "initial/EnzoInitialInclinedWave.hpp"
+#include "initial/EnzoInitialMusic.hpp"
+#include "initial/EnzoInitialPm.hpp"
+#include "initial/EnzoInitialPpmlTest.hpp"
+#include "initial/EnzoInitialSedovArray2.hpp"
+#include "initial/EnzoInitialSedovArray3.hpp"
+#include "initial/EnzoInitialSedovRandom.hpp"
+#include "initial/EnzoInitialShockTube.hpp"
+#include "initial/EnzoInitialSoup.hpp"
+#include "initial/EnzoInitialTurbulence.hpp"
+#include "initial/EnzoInitialIsolatedGalaxy.hpp"
+#include "initial/EnzoInitialBurkertBodenheimer.hpp"
+#include "initial/EnzoInitialMergeSinksTest.hpp"
+#include "initial/EnzoInitialAccretionTest.hpp"
+#include "initial/EnzoInitialShuCollapse.hpp"
+#include "initial/EnzoInitialBBTest.hpp"
 
 #include "enzo_EnzoRefineShock.hpp"
 #include "enzo_EnzoRefineParticleMass.hpp"
