+++ src/Enzo/initial/CMakeLists.txt
@@ -1,0 +1,19 @@
+# See LICENSE_CELLO file for license and copyright information
+
+# add source files related to problem-initialization to the enzo target
+#
+# in the future, we may want to slightly refactor the files in this
+# subdirectory so that they can be compiled into their own subtarget
+
+# Get the list of source files in this directory
+# - we do this using GLOB patterns. This approach is not recommended by the
+#   authors of CMake (their recommendation is to explicitly list all files that
+#   must be installed).
+# - Some of the disadvantages of this approach are mitigated by inclusion of
+#   the CONFIGURE_DEPENDS flag.
+# - See the CMake Primer section of the developer documentation for more details
+file(GLOB LOCAL_SRC_FILES CONFIGURE_DEPENDS
+  *.cpp *.hpp *F
+)
+
+target_sources(enzo PRIVATE ${LOCAL_SRC_FILES})
