+++ src/Enzo/initial/EnzoInitialImplosion2.cpp
