+++ src/Enzo/initial/EnzoInitialIsolatedGalaxy.cpp
