+++ src/Enzo/initial/EnzoInitialIsolatedGalaxy.hpp
